requirements.txt
Bottleneck==1.3.5
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.5
cvbase==0.5.5
cycler==0.11.0
dpcpp-cpp-rt==2023.0.0
einops==0.6.0
filelock==3.10.0
fonttools==4.25.0
huggingface-hub==0.13.2
idna==3.4
importlib-resources==5.12.0
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mkl==2023.0.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
networkx==3.0
numexpr==2.8.4
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.4.4
Pillow==9.4.0
ply==3.11
protobuf==3.20.3
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytorch-pretrained-vit==0.0.7
pytz==2022.7
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
sip==6.6.2
six==1.16.0
sympy==1.11.1
tasklib==2.5.1
tbb==2021.8.0
tensorboardX==2.6
terminaltables==3.1.10
timm==0.6.12
toml==0.10.2
torch
torchaudio
torchvision
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
urwid==2.1.2
wincertstore==0.2
zipp==3.15.0
